Unify Riverpod state management: drop wrapper classes and codegen tooling
- Delete OfflineApi and MapApi wrapper classes; the notifier is now the
  public API. Consumers call ref.read(provider.notifier).method directly.
- Move OfflineRegionsNotifier and helpers from offline_regions/api.dart
  into data/offline_regions_notifier.dart; api.dart is now a thin
  re-export facade. Fold coord computation into createRegion so callers
  pass bounds rather than precomputed coords.
- Add createTileProvider as a method on OfflineRegionsNotifier; switch
  to ref.read with a discarded ref.watch in tile_registry to keep
  reactivity on the calling widget rather than the wrapper provider.
- Reuse the existing top-level animatedMapMove in
  core/widgets/map/controller/map_utility.dart for the two callsites in
  main_map_page (was duplicated inside MapApi).
- Replace inline NotifierProvider<LoadingNotifier, bool> in login and
  register screens with plain setState; pass bool plus start/complete
  callbacks to the view widgets. Delete the now-unused LoadingNotifier
  from core/theme/utils.dart.
- Remove riverpod_annotation and build_runner from pubspec.yaml; the
  codebase has zero @riverpod annotations and zero .g.dart files.
- Update architecture doc: drop the wrapper-class recommendation, codify
  the canonical pattern (provider declared next to its notifier in data,
  re-exported from api.dart), document the deliberate seeded-AsyncValue
  pattern, and add a no-codegen rule.
