+++ lib/core/theme/utils.dart
@@ -1,12 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_fonts/google_fonts.dart';
-
-class LoadingNotifier extends Notifier<bool> {
-  @override
-  bool build() => false;
-  void set(bool value) => state = value;
-}
 
 TextTheme createTextTheme(
     BuildContext context, String bodyFontString, String displayFontString) {
