+++ lib/features/auth/widgets/login_screen.dart
@@ -3,7 +3,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../data/auth_providers.dart';
-import 'package:turbo/core/theme/utils.dart';
 import './register_screen.dart';
 
 import 'login_view_desktop.dart';
@@ -47,8 +46,8 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
   final _emailController = TextEditingController();
   final _passwordController = TextEditingController();
 
-  final _isLoading = NotifierProvider<LoadingNotifier, bool>(LoadingNotifier.new);
-  final _isGoogleLoading = NotifierProvider<LoadingNotifier, bool>(LoadingNotifier.new);
+  bool _isLoading = false;
+  bool _isGoogleLoading = false;
 
   @override
   void initState() {
@@ -81,7 +80,7 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
   Future<void> _login() async {
     FocusScope.of(context).unfocus();
     if (_formKey.currentState?.validate() ?? false) {
-      ref.read(_isLoading.notifier).set(true);
+      setState(() => _isLoading = true);
       try {
         await ref.read(authStateProvider.notifier).login(
           _emailController.text.trim(),
@@ -89,7 +88,7 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
         );
       } finally {
         if (mounted) {
-          ref.read(_isLoading.notifier).set(false);
+          setState(() => _isLoading = false);
         }
       }
     }
@@ -100,6 +99,14 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
       Navigator.of(context).pop();
     }
     RegisterScreen.show(context);
+  }
+
+  void _onGoogleSignInStarted() {
+    setState(() => _isGoogleLoading = true);
+  }
+
+  void _onGoogleSignInCompleted() {
+    if (mounted) setState(() => _isGoogleLoading = false);
   }
 
   @override
@@ -113,8 +120,10 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
         passwordController: _passwordController,
         onLogin: _login,
         onNavigateToRegister: _navigateToRegister,
-        isLoadingProvider: _isLoading,
-        isGoogleLoadingProvider: _isGoogleLoading,
+        isLoading: _isLoading,
+        isGoogleLoading: _isGoogleLoading,
+        onGoogleSignInStarted: _onGoogleSignInStarted,
+        onGoogleSignInCompleted: _onGoogleSignInCompleted,
       );
     } else {
       return LoginViewMobile(
@@ -123,8 +132,10 @@ class _LoginScreenState extends ConsumerState<LoginScreen> {
         passwordController: _passwordController,
         onLogin: _login,
         onNavigateToRegister: _navigateToRegister,
-        isLoadingProvider: _isLoading,
-        isGoogleLoadingProvider: _isGoogleLoading,
+        isLoading: _isLoading,
+        isGoogleLoading: _isGoogleLoading,
+        onGoogleSignInStarted: _onGoogleSignInStarted,
+        onGoogleSignInCompleted: _onGoogleSignInCompleted,
       );
     }
   }
