+++ lib/features/auth/widgets/login_view_mobile.dart
@@ -9,7 +9,6 @@ import './auth_text_field.dart';
 import './google_sign_in_button.dart';
 import './password_field.dart';
 import 'package:turbo/core/widgets/buttons/primary_button.dart';
-import 'package:turbo/core/theme/utils.dart';
 
 class LoginViewMobile extends ConsumerWidget {
   final GlobalKey<FormState> formKey;
@@ -17,8 +16,10 @@ class LoginViewMobile extends ConsumerWidget {
   final TextEditingController passwordController;
   final Future<void> Function() onLogin;
   final VoidCallback onNavigateToRegister;
-  final NotifierProvider<LoadingNotifier, bool> isLoadingProvider;
-  final NotifierProvider<LoadingNotifier, bool> isGoogleLoadingProvider;
+  final bool isLoading;
+  final bool isGoogleLoading;
+  final VoidCallback onGoogleSignInStarted;
+  final VoidCallback onGoogleSignInCompleted;
 
   const LoginViewMobile({
     super.key,
@@ -27,8 +28,10 @@ class LoginViewMobile extends ConsumerWidget {
     required this.passwordController,
     required this.onLogin,
     required this.onNavigateToRegister,
-    required this.isLoadingProvider,
-    required this.isGoogleLoadingProvider,
+    required this.isLoading,
+    required this.isGoogleLoading,
+    required this.onGoogleSignInStarted,
+    required this.onGoogleSignInCompleted,
   });
 
   @override
@@ -83,15 +86,15 @@ class LoginViewMobile extends ConsumerWidget {
                     PrimaryButton(
                       text: l10n.signIn,
                       onPressed: onLogin,
-                      isLoading: ref.watch(isLoadingProvider),
+                      isLoading: isLoading,
                     ),
                     const SizedBox(height: 24),
                     AuthDivider(text: l10n.or),
                     const SizedBox(height: 24),
                     GoogleSignInButton(
-                      isLoading: ref.watch(isGoogleLoadingProvider),
-                      onSignInStarted: () => ref.read(isGoogleLoadingProvider.notifier).set(true),
-                      onSignInCompleted: () => ref.read(isGoogleLoadingProvider.notifier).set(false),
+                      isLoading: isGoogleLoading,
+                      onSignInStarted: onGoogleSignInStarted,
+                      onSignInCompleted: onGoogleSignInCompleted,
                     ),
                     const SizedBox(height: 24),
                     AuthFooterLink(
