+++ lib/features/auth/widgets/register_view_mobile.dart
@@ -9,7 +9,6 @@ import './auth_text_field.dart';
 import './google_sign_in_button.dart';
 import './password_field.dart';
 import 'package:turbo/core/widgets/buttons/primary_button.dart';
-import 'package:turbo/core/theme/utils.dart';
 
 class RegisterViewMobile extends ConsumerWidget {
   final GlobalKey<FormState> formKey;
@@ -17,18 +16,22 @@ class RegisterViewMobile extends ConsumerWidget {
   final TextEditingController passwordController;
   final Future<void> Function() onRegister;
   final VoidCallback onNavigateToLogin;
-  final NotifierProvider<LoadingNotifier, bool> isLoadingProvider;
-  final NotifierProvider<LoadingNotifier, bool> isGoogleLoadingProvider;
+  final bool isLoading;
+  final bool isGoogleLoading;
+  final VoidCallback onGoogleSignInStarted;
+  final VoidCallback onGoogleSignInCompleted;
 
-  const RegisterViewMobile( {
+  const RegisterViewMobile({
     super.key,
     required this.formKey,
     required this.emailController,
     required this.passwordController,
     required this.onRegister,
     required this.onNavigateToLogin,
-    required this.isLoadingProvider,
-    required this.isGoogleLoadingProvider,
+    required this.isLoading,
+    required this.isGoogleLoading,
+    required this.onGoogleSignInStarted,
+    required this.onGoogleSignInCompleted,
   });
 
   @override
@@ -86,15 +89,15 @@ class RegisterViewMobile extends ConsumerWidget {
                     PrimaryButton(
                       text: l10n.createAccount,
                       onPressed: onRegister,
-                      isLoading: ref.watch(isLoadingProvider),
+                      isLoading: isLoading,
                     ),
                     const SizedBox(height: 24),
                     AuthDivider(text: l10n.or),
                     const SizedBox(height: 24),
                     GoogleSignInButton(
-                      isLoading: ref.watch(isGoogleLoadingProvider),
-                      onSignInStarted: () => ref.read(isGoogleLoadingProvider.notifier).set(true),
-                      onSignInCompleted: () => ref.read(isGoogleLoadingProvider.notifier).set(false),
+                      isLoading: isGoogleLoading,
+                      onSignInStarted: onGoogleSignInStarted,
+                      onSignInCompleted: onGoogleSignInCompleted,
                     ),
                     const SizedBox(height: 24),
                     Text(
