+++ pubspec.yaml
@@ -51,7 +51,6 @@ dependencies:
   geolocator: ^14.0.2
   flutter_svg: ^2.2.3
   flutter_riverpod: ^3.2.1
-  riverpod_annotation: ^4.0.2
   google_fonts: ^8.0.2
   shared_preferences: ^2.5.4
   share_plus: ^10.1.4
@@ -76,7 +75,6 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.11.1
   custom_lint: ^0.8.1
   http_mock_adapter: ^0.6.1
   shelf: ^1.4.2
